requirements.txt
crewai==0.28.8
groq==0.4.1
pandas==2.1.4
pydantic==2.5.3
pyyaml==6.0.1
rich==13.7.0
loguru==0.7.2
python-dotenv==1.0.0

